WIP Wee update in Bob Tutorial NB. Websocket showcase
* Little more structure to Bob's tutorial Notebook. Still needs work
* Check out the examples in the Websocket notebooks in the Bob folder
** This is a minimal-ish implementation of a Websocket server instead of
http that can handle multiple client connections to solve the current
webhook port-in-use problem.
** It already demonstrates running multiple clients on the same
endpoint. Here 127.0.0.1:8090/ws.
** I reckon we can make this quite elegant by re-writing this further
and using routes like /ws/{some-unique-wallet-id}/the-webhook-handle to
send specific webhooks only to the actual recepient.
** For the previous point could even use authentication with that (not
sure that's required actually because the agent should already be only
able to connect when having the right credential?!) TODO: investigate
this point.